requirements.txt
pyjwt==2.12.1
cryptography==41.0.7
click==8.1.7
rich==13.7.0
reportlab==4.0.9
